requirements.txt
streamlit
# transformers
pdfplumber
# tensorflow
# tf-keras
google.generativeai
python-dotenv